[요청]
PR Table Schema Changed

[확인]
고객관련 스키마 수정확인
키/몸무게 integer -> float

+++ src/main/java/com/dogfeetbirdfeet/ivorypilatesbackend/dto/schema/CusCons.java
@@ -11,8 +11,8 @@ public class CusCons {
 
 	private String cusId;
 	private String mstId;
-	private Integer height;
-	private Integer weight;
+	private Float height;
+	private Float weight;
 	private String disease;
 	private String surHist;
 	private String bodyCheckImg;
